feat(handleoption.js): add a sort API that supports sorting by value or percent in (asc) or (desc)
Added a sorting API that supports sorting by value or percent in asc or desc order.
