+++ public/option/RankProcessChart/RankProcessChart-3.js
@@ -7,24 +7,78 @@ const option = {
     // padding控制图表距离容器的上、右、下、左padding值
     padding: [16, 16, 0, 16], 
     
-    // 自定义图表信息 - 可自定义列标题
-    titleName: '部门',
-    valueName: '业绩(万元)',
-    percentName: '达成率',
+    titleName: '团队成员',
+    valueName: '工作时长',
+    percentName: '效率',
     
-    // Tooltip 配置
+    // 排序配置
+    sort: {
+        field: 'value',
+        order: 'desc'
+    },
+    
+    // Tooltip 自定义格式化配置
     tooltip: {
         show: true,
-        formatter: null
+        formatter: function(params) {
+            const { name, value, percent, color, content } = params;
+            
+            return `
+                <div style="padding: 8px; min-width: 200px;">
+                    <div style="font-weight: bold; font-size: 16px; margin-bottom: 8px; color: ${color};">
+                        ${name}
+                    </div>
+                    <div style="margin-bottom: 4px;">
+                        <span style="display: inline-block; width: 10px; height: 10px; background-color: ${color}; border-radius: 50%; margin-right: 8px;"></span>
+                        工作时长: <strong>${value}小时</strong>
+                    </div>
+                    <div style="margin-bottom: 4px;">
+                        效率指数: <strong>${percent}%</strong>
+                    </div>
+                    <div style="border-top: 1px solid #eee; padding-top: 6px; margin-top: 6px; font-size: 12px; color: #666;">
+                        ${content}
+                    </div>
+                    <div style="text-align: right; font-size: 11px; color: #999; margin-top: 4px;">
+                        数据更新时间: ${new Date().toLocaleTimeString()}
+                    </div>
+                </div>
+            `;
+        }
     },
     
     // 数据
     data: [
-        { name: '销售一部', value: 1250, percent: 95, content: '销售一部业绩1250万元，达成率95%，表现优异' },
-        { name: '销售二部', value: 980, percent: 78, content: '销售二部业绩980万元，达成率78%，进展良好' },
-        { name: '销售三部', value: 1420, percent: 88, content: '销售三部业绩1420万元，达成率88%，业绩领先' },
-        { name: '销售四部', value: 760, percent: 65, content: '销售四部业绩760万元，达成率65%，需要提升' },
-        { name: '销售五部', value: 1100, percent: 82, content: '销售五部业绩1100万元，达成率82%，稳步增长' }
+        { 
+            name: '张三', 
+            value: 42, 
+            percent: 95, 
+            content: '高级前端工程师，本周完成了用户界面优化和组件重构工作，表现优异' 
+        },
+        { 
+            name: '李四', 
+            value: 38, 
+            percent: 88, 
+            content: '后端开发工程师，负责API接口开发和数据库优化，进展顺利' 
+        },
+        { 
+            name: '王五', 
+            value: 35, 
+            percent: 82, 
+            content: '测试工程师，完成了自动化测试脚本编写和功能测试，质量保证到位' 
+        },
+        { 
+            name: '赵六', 
+            value: 40, 
+            percent: 90, 
+            content: '产品经理，负责需求分析和产品设计，沟通协调能力强' 
+        },
+        { 
+            name: '钱七', 
+            value: 33, 
+            percent: 75, 
+            content: 'UI设计师，完成了界面设计和交互原型，创意能力突出' 
+        }
     ],
     
 };
+
