--- public/option/RankProcessChart/RankProcessChart-7.js
@@ -1,72 +1,0 @@
-const option = {
-    // 图表名称
-    name: 'RankProcessChart',
-    // 主题,默认值'hdesign-light'
-    theme: 'hdesign-light',
-
-    // padding控制图表距离容器的上、右、下、左padding值
-    padding: [16, 16, 0, 16], 
-    
-    titleName: '项目全称',
-    valueName: '预算(万元)',
-    percentName: '执行率',
-    
-    // Tooltip 配置
-    tooltip: {
-        show: true,
-        formatter: null
-    },
-    
-    // 数据
-    data: [
-        { 
-            name: '企业级云原生微服务架构数字化转型平台建设项目', 
-            value: 2850, 
-            percent: 92, 
-            content: '企业级云原生微服务架构数字化转型平台建设项目预算2850万元，执行率92%，项目进展顺利' 
-        },
-        { 
-            name: '基于人工智能的智能客服系统与数据分析平台集成方案', 
-            value: 1980, 
-            percent: 78, 
-            content: '基于人工智能的智能客服系统与数据分析平台集成方案预算1980万元，执行率78%，AI算法优化中' 
-        },
-        { 
-            name: '移动端跨平台应用开发与用户体验优化升级项目', 
-            value: 3420, 
-            percent: 85, 
-            content: '移动端跨平台应用开发与用户体验优化升级项目预算3420万元，执行率85%，界面设计完成' 
-        },
-        { 
-            name: '大数据实时处理与商业智能分析决策支持系统', 
-            value: 4150, 
-            percent: 88, 
-            content: '大数据实时处理与商业智能分析决策支持系统预算4150万元，执行率88%，数据仓库建设完成' 
-        },
-        { 
-            name: '区块链技术在企业供应链管理与溯源系统中的应用', 
-            value: 2680, 
-            percent: 95, 
-            content: '区块链技术在企业供应链管理与溯源系统中的应用预算2680万元，执行率95%，智能合约开发完成' 
-        },
-        { 
-            name: '物联网设备管理与边缘计算平台建设项目', 
-            value: 1750, 
-            percent: 73, 
-            content: '物联网设备管理与边缘计算平台建设项目预算1750万元，执行率73%，设备接入协议开发中' 
-        },
-        { 
-            name: '企业级安全防护与合规性管理综合解决方案', 
-            value: 3280, 
-            percent: 81, 
-            content: '企业级安全防护与合规性管理综合解决方案预算3280万元，执行率81%，安全策略配置完成' 
-        },
-        { 
-            name: '基于机器学习的用户行为分析与个性化推荐引擎', 
-            value: 1450, 
-            percent: 90, 
-            content: '基于机器学习的用户行为分析与个性化推荐引擎预算1450万元，执行率90%，推荐算法优化完成' 
-        }
-    ],
-    
-};
